Remove all dependencies from dot net framework. (see #15)
