+++ Source/BSN.Commons.Orm.EntityFramework/RepositoryBase.cs
@@ -4,15 +4,17 @@ using System.Data.Entity;
 using System.Linq;
 using System.Linq.Expressions;
 
-namespace Commons.Infrastructure
+namespace Commons.Orm.EntityFramework
 {
+    using Commons.Infrastructure;
+
     public abstract partial class RepositoryBase<T> : IRepository<T> where T : class
     {
         private DbContext _dataContext;
         private readonly IUnitOfWork _unitOfWork;
         protected readonly DbSet<T> dbSet;
 
-        protected DbContext DataContext => _dataContext ?? (_dataContext = DatabaseFactory.Get());
+        protected DbContext DataContext => _dataContext ?? (_dataContext = (DbContext)DatabaseFactory.Get());
         protected IDatabaseFactory DatabaseFactory { get; private set; }
 
 
@@ -165,4 +167,76 @@ namespace Commons.Infrastructure
             return dbSet.Where(where).FirstOrDefault();
         }
     }
+
+	public abstract partial class RepositoryBase<T> where T : class
+	{
+		private class UpdateConfig : IUpdateConfig<T> 
+		{
+			public bool IncludeAllPropertiesEnabled
+			{
+				get
+				{
+					return _includeAllPropertiesEnabled;
+				}
+				private set
+				{
+					_includeAllPropertiesEnabled = value;
+					if (value == true)
+						_autoDetectChangedPropertiesEnabled = false;
+				}
+			}
+
+			public bool AutoDetectChangedPropertiesEnabled
+			{
+				get
+				{
+					return _autoDetectChangedPropertiesEnabled;
+				}
+				private set
+				{
+					_autoDetectChangedPropertiesEnabled = value;
+					if (value == true)
+						_includeAllPropertiesEnabled = false;
+				}
+			}
+
+			public IList<string> PropertyNames { get; }
+
+			public UpdateConfig()
+			{
+				PropertyNames = new List<string>();
+			}
+
+			public void IncludeAllProperties()
+			{
+				IncludeAllPropertiesEnabled = true;
+			}
+
+			public void AutoDetectChangedProperties()
+			{
+				AutoDetectChangedPropertiesEnabled = true;
+			}
+
+			public IUpdateConfig<T> IncludeProperty<TProperty>(Expression<Func<T, TProperty>> propertyAccessExpression)
+			{
+				PropertyNames.Add(
+					(propertyAccessExpression.Body as MemberExpression)?.Member.Name ??
+						throw new ArgumentException(nameof(propertyAccessExpression))
+				);
+
+				AutoDetectChangedPropertiesEnabled = false;
+				IncludeAllPropertiesEnabled = false;
+				return this;
+			}
+
+			#region Fields
+
+			private bool _includeAllPropertiesEnabled;
+
+			private bool _autoDetectChangedPropertiesEnabled;
+
+			#endregion
+		}
+    }
+
 }
