+++ Source/BSN.Commons.Orm.EntityFramework/Users/Role.cs
