+++ Source/BSN.Commons.Orm.EntityFramework/Users/User.cs
