+++ Source/BSN.Commons.Orm.EntityFramework/Users/UserRole.cs
