+++ Source/BSN.Commons.Orm.EntityFramework/packages.config
@@ -1,0 +1,6 @@
+﻿<?xml version="1.0" encoding="utf-8"?>
+<packages>
+  <package id="EntityFramework" version="6.2.0" targetFramework="net48" />
+  <package id="Microsoft.AspNet.Identity.Core" version="2.2.3" targetFramework="net48" />
+  <package id="Microsoft.AspNet.Identity.EntityFramework" version="2.2.3" targetFramework="net48" />
+</packages>
