+++ Source/BSN.Commons/Infrastructure/IDatabaseFactory.cs
@@ -1,10 +1,9 @@
 ﻿using System;
-using System.Data.Entity;
 
 namespace Commons.Infrastructure
 {
     public interface IDatabaseFactory : IDisposable
     {
-        DbContext Get();
+        IDbContext Get();
     }
 }
