+++ Source/BSN.Commons/Infrastructure/IDbContext.cs
@@ -1,0 +1,13 @@
+﻿namespace Commons.Infrastructure
+{
+	public interface IDbSet<TEntity> where TEntity : class
+	{
+
+	}
+
+	public interface IDbContext
+	{
+		void SaveChanges();
+		IDbSet<TEntity> Set<TEntity>() where TEntity : class;
+	}
+}
